Unify merge commit message
