+++ src/database/mod.rs
@@ -271,7 +271,7 @@ impl FromStr for DelegatedPermission {
 }
 
 /// Status of a GitHub build.
-#[derive(Debug, PartialEq, sqlx::Type)]
+#[derive(Debug, Clone, PartialEq, sqlx::Type)]
 #[sqlx(type_name = "TEXT")]
 #[sqlx(rename_all = "lowercase")]
 pub enum BuildStatus {
@@ -300,7 +300,7 @@ impl Display for BuildStatus {
 }
 
 /// Represents a single (merged) commit.
-#[derive(Debug, sqlx::Type)]
+#[derive(Debug, Clone, sqlx::Type)]
 #[sqlx(type_name = "build")]
 pub struct BuildModel {
     pub id: PrimaryKey,
@@ -320,7 +320,7 @@ pub struct BuildModel {
 }
 
 /// Represents a pull request.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct PullRequestModel {
     pub id: PrimaryKey,
     /// The GitHub repository this PR belongs to.
